+++ front_end/app/static/css/styleProfile.css
@@ -118,17 +118,54 @@ line{
     border-bottom: 5px solid #757575;
 }
 
-.AdminPI{
+.AdminPIContainer{
+    justify-content: center;
     position: relative;
     top: 100px;
     margin-left: auto;
     margin-right: auto;
-    height: 200px;
-    width: 300px;
+    height: 250px;
+    width: 330px;
     padding: 5px 5px;
-    background-color: #f0f0f0;
+    background-color: #ffffff;
 }
 
+.AdminPIContainer .AdminPI{
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+}
+
+.AdminPIContainer .buttonPI{
+    margin-top: 10px;
+    margin-left: auto;
+    margin-right: auto;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: #ffffff;
+    width: 75px;
+    height: 25px;
+    text-align: center;
+    cursor: pointer;
+}
+
+.AdminPIContainer .buttonPI:hover{
+    background-color: #dadada;
+}
+
+.AdminPIContainer .buttonPI:active{
+    background-color: #a1a1a1;
+}
+
+
 .AdminTitle{
+    text-align: center;
+}
+
+.ErrorMessage{
+    color: rgb(218, 0, 0);
+    text-size-adjust: 20px;
     text-align: center;
 }
